+++ ProjectAscension.pro
@@ -21,20 +21,24 @@ HEADERS += \
     Source/Game.h \
     Source/TabLabel.h \
     Source/Library.h \
-    Source/Browser.h
-
+    Source/Browser.h \
+    Source/MainPanel.h
 
 SOURCES += \
     Source/Main.cpp \
     Source/TabLabel.cpp \
     Source/Library.cpp \
-    Source/Browser.cpp
+    Source/Browser.cpp \
+    Source/MainPanel.cpp
 
 win32 {
     HEADERS += \
+        Source/WinMainPanel.h \
         Source/WinPanel.h \
         Source/BorderlessWindow.h
+
     SOURCES += \
+        Source/WinMainPanel.cpp \
         Source/WinPanel.cpp \
         Source/BorderlessWindow.cpp
 }
